Draw points from a dense flat vector
Now I copy the scaled data points to a vector containing a single
float value representing every pixel in the sketch window.
I changed and deleted a lot of things in the process of getting
this to a working state. I definitely should have done more unit
testing on small functions to make debugging more manageable,
but I got there.

Added one test. It's a start...