Currently producing art
Now that the interpolation algorithm runs in parallel in
a reasonable amount of time, it's plain to see that this is bad
at producing real altitude/elevation plots, but it makes some cool
art. So, I'm going to play around with that aspect a bit.
Maybe I'll start a new branch for that.